Update struct io_uring_files_update struct
Made a change to ensure 32-bit vs 64-bit compatability. This isn't in
a released kernel yet, so make the change now.

+++ src/register.c
@@ -48,7 +48,7 @@ int io_uring_register_files_update(struct io_uring *ring, unsigned off,
 {
 	struct io_uring_files_update up = {
 		.offset	= off,
-		.fds	= files,
+		.fds	= (unsigned long) files,
 	};
 	int ret;
